requirements.txt
prompt-toolkit
GitPython
argparse